LPD-20797 reorganize to their proper classes
