+++ modules/apps/change-tracking/change-tracking-test/src/testIntegration/java/com/liferay/change/tracking/internal/performance/test/CTCollectionLocalServicePerformanceTest.java
@@ -10,34 +10,19 @@ import com.liferay.change.tracking.model.CTCollection;
 import com.liferay.change.tracking.service.CTCollectionLocalService;
 import com.liferay.change.tracking.service.CTProcessLocalService;
 import com.liferay.change.tracking.test.util.PerformanceTestTimer;
-import com.liferay.dynamic.data.mapping.model.DDMForm;
-import com.liferay.dynamic.data.mapping.model.DDMStructure;
-import com.liferay.dynamic.data.mapping.model.DDMTemplate;
-import com.liferay.dynamic.data.mapping.test.util.DDMFormTestUtil;
-import com.liferay.dynamic.data.mapping.test.util.DDMFormValuesTestUtil;
-import com.liferay.dynamic.data.mapping.test.util.DDMStructureTestUtil;
-import com.liferay.dynamic.data.mapping.test.util.DDMTemplateTestUtil;
-import com.liferay.journal.constants.JournalArticleConstants;
-import com.liferay.journal.model.JournalArticle;
-import com.liferay.journal.model.JournalFolder;
-import com.liferay.journal.service.JournalArticleLocalService;
-import com.liferay.journal.test.util.JournalTestUtil;
 import com.liferay.layout.test.util.LayoutTestUtil;
 import com.liferay.petra.lang.SafeCloseable;
 import com.liferay.portal.kernel.change.tracking.CTCollectionThreadLocal;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.GroupConstants;
-import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DataGuard;
-import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
@@ -45,7 +30,6 @@ import com.liferay.site.initializer.SiteInitializer;
 import com.liferay.site.initializer.SiteInitializerRegistry;
 
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
@@ -57,7 +41,7 @@ import org.junit.runner.RunWith;
  */
 @DataGuard(scope = DataGuard.Scope.NONE)
 @RunWith(Arquillian.class)
-public class CTRowUtilPerformanceTest {
+public class CTCollectionLocalServicePerformanceTest {
 
 	@ClassRule
 	@Rule
@@ -84,81 +68,6 @@ public class CTRowUtilPerformanceTest {
 		_ctCollectionLocalService.deleteCTCollection(_ctCollection2);
 
 		GroupTestUtil.deleteGroup(_group);
-	}
-
-	@Test
-	public void testAddCTProcessDDMForm() throws Exception {
-		String formFieldName = RandomTestUtil.randomString();
-
-		DDMForm ddmForm = DDMFormTestUtil.createDDMForm(
-			RandomTestUtil.randomString(), RandomTestUtil.randomString());
-
-		try (SafeCloseable safeCloseable =
-				CTCollectionThreadLocal.setCTCollectionIdWithSafeCloseable(
-					_ctCollection1.getCtCollectionId())) {
-
-			DDMFormTestUtil.addTextDDMFormFields(ddmForm, formFieldName);
-
-			DDMFormValuesTestUtil.createDDMFormValuesWithRandomValues(ddmForm);
-		}
-
-		DDMFormValuesTestUtil.createDDMFormValuesWithRandomValues(ddmForm);
-
-		DDMFormTestUtil.addTextDDMFormFields(ddmForm, formFieldName);
-
-		try (PerformanceTestTimer performanceTestTimer =
-				new PerformanceTestTimer(2000)) {
-
-			_ctProcessLocalService.addCTProcess(
-				_ctCollection1.getUserId(), _ctCollection1.getCtCollectionId());
-		}
-	}
-
-	@Test
-	public void testAddCTProcessJournalArticle() throws Exception {
-		JournalFolder journalFolder = JournalTestUtil.addFolder(
-			_group.getGroupId(), RandomTestUtil.randomString());
-
-		DDMStructure ddmStructure = DDMStructureTestUtil.addStructure(
-			_group.getGroupId(), JournalArticle.class.getName());
-
-		DDMTemplate ddmTemplate = DDMTemplateTestUtil.addTemplate(
-			_group.getGroupId(), ddmStructure.getStructureId(),
-			_portal.getClassNameId(JournalArticle.class));
-
-		JournalArticle journalArticle =
-			JournalTestUtil.addArticleWithXMLContent(
-				_group.getGroupId(), journalFolder.getFolderId(),
-				JournalArticleConstants.CLASS_NAME_ID_DEFAULT,
-				DDMStructureTestUtil.getSampleStructuredContent(),
-				ddmStructure.getStructureKey(), ddmTemplate.getTemplateKey());
-
-		JournalArticle ctJournalArticle = null;
-
-		try (SafeCloseable safeCloseable =
-				CTCollectionThreadLocal.setCTCollectionIdWithSafeCloseable(
-					_ctCollection1.getCtCollectionId())) {
-
-			ctJournalArticle = JournalTestUtil.updateArticle(journalArticle);
-		}
-
-		Assert.assertNotEquals(
-			journalArticle.getTitle(), ctJournalArticle.getTitle());
-
-		journalArticle = JournalTestUtil.updateArticle(journalArticle);
-
-		try (PerformanceTestTimer performanceTestTimer =
-				new PerformanceTestTimer(2000)) {
-
-			_ctProcessLocalService.addCTProcess(
-				_ctCollection1.getUserId(), _ctCollection1.getCtCollectionId());
-		}
-
-		journalArticle = _journalArticleLocalService.getLatestArticle(
-			journalArticle.getResourcePrimKey());
-
-		Assert.assertEquals(
-			journalArticle.getTitle(), ctJournalArticle.getTitle());
 	}
 
 	@Test
@@ -204,14 +113,14 @@ public class CTRowUtilPerformanceTest {
 		}
 
 		try (PerformanceTestTimer performanceTestTimer =
-				new PerformanceTestTimer(20000)) {
+				new PerformanceTestTimer(45000)) {
 
 			_ctProcessLocalService.addCTProcess(
 				_ctCollection2.getUserId(), _ctCollection2.getCtCollectionId());
 		}
 
 		try (PerformanceTestTimer performanceTestTimer =
-				new PerformanceTestTimer(1000)) {
+				new PerformanceTestTimer(10000)) {
 
 			_ctCollectionLocalService.checkConflicts(_ctCollection1);
 		}
@@ -238,23 +147,20 @@ public class CTRowUtilPerformanceTest {
 		}
 
 		try (PerformanceTestTimer performanceTestTimer =
-				new PerformanceTestTimer(2000)) {
+				new PerformanceTestTimer(10000)) {
 
 			_ctProcessLocalService.addCTProcess(
 				_ctCollection2.getUserId(), _ctCollection2.getCtCollectionId());
 		}
 
 		try (PerformanceTestTimer performanceTestTimer =
-				new PerformanceTestTimer(1000)) {
+				new PerformanceTestTimer(10000)) {
 
 			_ctCollectionLocalService.checkConflicts(_ctCollection1);
 		}
 	}
 
-	@DeleteAfterTestRun
 	private CTCollection _ctCollection1;
-
-	@DeleteAfterTestRun
 	private CTCollection _ctCollection2;
 
 	@Inject
@@ -263,17 +169,7 @@ public class CTRowUtilPerformanceTest {
 	@Inject
 	private CTProcessLocalService _ctProcessLocalService;
 
-	@DeleteAfterTestRun
 	private Group _group;
-
-	@Inject
-	private JournalArticleLocalService _journalArticleLocalService;
-
-	@Inject
-	private LayoutLocalService _layoutLocalService;
-
-	@Inject
-	private Portal _portal;
 
 	@Inject
 	private SiteInitializerRegistry _siteInitializerRegistry;
